+++ src/api/changelog.rs
@@ -40,8 +40,8 @@ mod tests {
         let changelog = Changelog { client };
 
         let params = ChangelogPersonParams {
-            origin_version: Some("33.1".to_string()),
-            current_version: Some("33.2".to_string()),
+            origin_version: Some("34.2".to_string()),
+            current_version: Some("35.0".to_string()),
             type_: Some("updated".to_string()),
             ..Default::default()
         };
@@ -62,8 +62,8 @@ mod tests {
         let changelog = Changelog { client };
 
         let params = ChangelogPersonParams {
-            origin_version: Some("33.1".to_string()),
-            current_version: Some("33.2".to_string()),
+            origin_version: Some("34.2".to_string()),
+            current_version: Some("35.0".to_string()),
             type_: Some("updated".to_string()),
             ..Default::default()
         };