+++ src/models/company.rs
@@ -156,7 +156,7 @@ pub struct SearchCompanyResponse {
 
 #[derive(Debug, Serialize, Deserialize)]
 pub struct CompanyError {
-    #[serde(rename = "type")]
+    #[serde(rename = "type", deserialize_with = "crate::models::common::deserialize_string_or_vec")]
     pub error_type: Vec<String>,
     pub message: String,
 }
@@ -292,6 +292,14 @@ pub struct EmployeeGrowthRate12MonthByCountry {
 }
 
 #[derive(Debug, Serialize, Deserialize)]
+pub struct TechnologiesUsed {
+    pub product_id: Option<String>,
+    pub name: Option<String>,
+    pub num_job_posting_mentions: Option<i32>,
+    pub num_resume_mentions: Option<i32>,
+}
+
+#[derive(Debug, Serialize, Deserialize)]
 pub struct CompanyResponse {
     /// See https://docs.peopledatalabs.com/docs/example-company-record for more information.
     pub status: Option<i32>,
@@ -382,6 +390,7 @@ pub struct CompanyResponse {
     pub funding_stages: Option<Vec<String>>,
     pub funding_details: Option<Vec<FundingDetails>>,
     pub likelihood: Option<i32>,
+    pub technologies_used: Option<Vec<TechnologiesUsed>>,
 }
 
 #[derive(Debug, Deserialize, Serialize)]